renamed sparse_table to range_minimum_query
