+++ src/data_structures/mod.rs
@@ -12,10 +12,10 @@ mod linked_list;
 mod postfix_evaluation;
 mod probabilistic;
 mod queue;
+mod range_minimum_query;
 mod rb_tree;
 mod segment_tree;
 mod segment_tree_recursive;
-mod sparse_table;
 mod stack_using_singly_linked_list;
 mod treap;
 mod trie;
@@ -38,10 +38,10 @@ pub use self::postfix_evaluation::evaluate_postfix;
 pub use self::probabilistic::bloom_filter;
 pub use self::probabilistic::count_min_sketch;
 pub use self::queue::Queue;
+pub use self::range_minimum_query::RangeMinimumQuery;
 pub use self::rb_tree::RBTree;
 pub use self::segment_tree::SegmentTree;
 pub use self::segment_tree_recursive::SegmentTree as SegmentTreeRecursive;
-pub use self::sparse_table::RangeMinimumQuery;
 pub use self::stack_using_singly_linked_list::Stack;
 pub use self::treap::Treap;
 pub use self::trie::Trie;
