+++ src/data_structures/range_minimum_query.rs
